Added fair play to Play command, disable shuffle when fp is enabled
