+++ src/commands/music/FairPlay.ts
@@ -1,5 +1,5 @@
 import {Command, type Lavamusic} from "../../structures/index";
-import {Track} from "lavalink-client";
+import {applyFairPlayToQueue} from '../../utils/functions/player';
 
 export default class FairPlay extends Command {
     constructor(client: Lavamusic) {
@@ -17,7 +17,7 @@ export default class FairPlay extends Command {
             vote: false,
             player: {
                 voice: true,
-                dj: false,
+                dj: true,
                 active: true,
                 djPerm: null,
             },
@@ -53,36 +53,7 @@ export default class FairPlay extends Command {
             embed.setDescription(ctx.locale('cmd.fairplay.messages.disabled')).setColor(this.client.color.main);
         } else {
             embed.setDescription(ctx.locale('cmd.fairplay.messages.enabled')).setColor(this.client.color.main);
-
-            const tracks = [...player.queue.tracks];
-            const requesterMap = new Map<string, any[]>();
-
-            // Group tracks by requester
-            for (const track of tracks) {
-                const requesterId = (track.requester as any).id
-                client.logger.log(`Requester ID: ${requesterId}`);
-                if (!requesterMap.has(requesterId)) {
-                    requesterMap.set(requesterId, []);
-                }
-                requesterMap.get(requesterId)!.push(track);
-            }
-
-            // Build fair queue
-            const fairQueue: Track[] = [];
-            while (fairQueue.length < tracks.length) {
-                for (const [, trackList] of requesterMap.entries()) {
-                    if (trackList.length > 0) {
-                        fairQueue.push(trackList.shift()!);
-                    }
-                }
-            }
-
-            // Reset queue
-            await player.queue.splice(0, player.queue.tracks.length);
-            for (const track of fairQueue) {
-                client.logger.log(`Adding track to fair queue: ${track.info.title}`);
-                player.queue.add(track);
-            }
+            await applyFairPlayToQueue(player);
         }
 
         await ctx.sendMessage({ embeds: [embed] });
